+++ libmpathpersist/mpath_updatepr.c
@@ -1,6 +1,7 @@
 #include <stdio.h>
 #include <unistd.h>
 #include <stdlib.h>
+#include <string.h>
 #include <stdarg.h>
 #include <fcntl.h>
 #include <sys/ioctl.h>
